+++ components/updater/api/src/main/java/com/flipperdevices/updater/model/SubGhzProvisioningModel.kt
@@ -1,0 +1,14 @@
+package com.flipperdevices.updater.model
+
+data class SubGhzProvisioningModel(
+    val countries: Map<String, List<SubGhzProvisioningBand>>,
+    val country: String,
+    val defaults: List<SubGhzProvisioningBand>
+)
+
+data class SubGhzProvisioningBand(
+    val dutyCycle: UInt,
+    val end: UInt,
+    val maxPower: Int,
+    val start: UInt
+)